GetRates method

+++ Assets/Scripts/UtilityRates.cs
@@ -1,18 +1,18 @@
 ﻿public class UtilityRates
 {
-    private double zipCode;
+    private int zipCode;
     private float electricityPerKWH;
     private float gasPerTherm;
     private float oilPerGallon;
     private float woodPerPound;
 
-    public float ZipCode { get => zipCode; }
+    public int ZipCode { get => zipCode; }
     public float ElectricityPerKWH { get => electricityPerKWH; }
     public float GasPerTherm { get => gasPerTherm; }
     public float OilPerGallon { get => oilPerGallon; }
     public float WoodPerPound { get => woodPerPound; }
 
-    public UtilityRates(float zipCode, float electricityPerKWH, float gasPerTherm, float oilPerGallon, float woodPerPound)
+    public UtilityRates(int zipCode, float electricityPerKWH, float gasPerTherm, float oilPerGallon, float woodPerPound)
     {
         this.zipCode = zipCode;
         this.electricityPerKWH = electricityPerKWH;
@@ -35,7 +35,7 @@
         var columns = row.Split(',');
         return new Rates()
         {
-            zipCode = float.Parse(columns[1]),
+            zipCode = int.Parse(columns[1]),
             electricityPerKWH = float.Parse(columns[2]),
             gasPerTherm = float.Parse(columns[3]),
             oilPerGallon = float.Parse(columns[4]),
